Improve memory performance of receive function
Previously a byte slice of size 65499 was created with each receive. This change
reuses the same slice. This in turn means introduces that the data in listener
is not thread-safe. This commit purposefully does not introduce a mutex as one
should not define multiple listeners for receiving. We want listeners within the
same program to receive the same multicast-packet at most once.