Update API to use context.Context for timeouts
Simplifies use of the LCM Transmitter/Receiver when there is already a
context with a deadline associated with it.

Also simplifies resetting of the deadline since it is automatically reset
on every call, if a deadline is not set.

Updates unit tests to use mocked UDP readers and writers.

Updates integration tests with the new API.

100% test coverage.